requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
bleach==6.1.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
cytoolz==0.12.3
datasets==2.19.1
dill==0.3.8
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.3
eth-utils==4.1.1
eth_abi==5.1.0
exceptiongroup==1.2.1
filelock==3.14.0
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.3.1
hexbytes==0.3.1
huggingface-hub==0.23.2
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
MarkupSafe==2.1.5
multidict==6.0.5
multiprocess==0.70.16
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parsimonious==0.10.0
pluggy==1.5.0
protobuf==5.27.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycryptodome==3.20.0
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
rlp==4.0.1
rpds-py==0.18.1
six==1.16.0
story-protocol-python-sdk==0.2.1
text-unidecode==1.3
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
web3==6.19.0
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
xxhash==3.4.1
yarl==1.9.4
